fix: allow mqtt client to reconnect if it gets disconnected (#977)
MQTT client reconnections were disabled for seemingly no reason.
Combined with the fact that the SocketClient's various error and
disconnection events were completely ignored, a random disconnection
before the expected final message was received would almost certainly
cause the CLI to essentially hang, doing nothing until the failsafe
timeout (currently 10m) hit.

There is an additional edge case which is not covered by this
change. Even now that reconnections are enabled, in the unlucky case
that the final message is distributed when our client is not connected
(i.e. in the process waiting to reconnect, or actively reconnecting),
then we'll still miss it and hang the same way. To reduce the odds
of this edge condition happening, the reconnect wait timeout was
reduced to 100ms from the default of 1000ms.