feat: compress large JSON requests
When you deploy or test a project, a potentially very large JSON payload is
sent to the API server. In large projects, the time to upload the payload can
be quite significant (depending on connection speed). This PR introduces gzip
compression to payloads sent to relevant endpoints, which should make the
payload easier to upload.

Additionally, those large payload were stringified fully in memory, causing
unnecessarily high memory usage. This PR changes the payload to be
stringified and streamed with json-stream-stringify instead, which should
decrease memory usage. This change only applies to requests that are now
also getting compressed.