feat: support wildcard exports when parsing dependency tree [sc-22632] (#1003)
* feat: add a hidden flag to output the project bundle for debug purposes

The new flag, `--debug-bundle`, outputs the project bundle payload as-is
after the project has been parsed. The bundle is output to a file, by default
`./debug-bundle.json`. A different file can be chosen with the new
`--debug-bundle-output-file` flag, which is also hidden.

* feat: support wildcard exports when parsing code for dependencies

Previously only named exports were supported.

* feat: use wildcard export in tests

Additionally, to make sure that not all files in the tree are getting
included automatically (which would have invalidated the wildcard export
tests), an unreachable file that should not appear in the list of
dependencies was added to the relevant examples.