feat: resolve declaration files (e.g. `.d.ts`) as a last resort
It seems that some users are using declaration files as shared header files
and there is nothing else to resolve. During normal transpilation, any such
imports would get removed and the declaration file would not be part of the
dist bundle. Since we transpile on the runner, it would be nice to have the
file available there even though it's not technically even required. More
importantly however this change makes the parser not complain about missing
dependencies if it encounters imports that can only be resolved to a
declaration file.