feat: allow multiple references to the same resource
It is now possible to use AlertChannel.fromId() and PrivateLocation.fromId()
multiple times without the project complaining about a conflicting logicalId.

Currently the only non-member resources are AlertChannelRef (renamed from
AlertChannelWrapper) and PrivateLocationRef (renamed from
PrivateLocationWrapper). Previously they were not exported, but now are, so
that they can be added to type definitions.

This also fixes an issue where PrivateLocation.fromId() could not actually
be used with Checks and CheckGroups because the privateLocations array did
not include the previously non-exposed PrivateLocationWrapper (now
exposed as PrivateLocationRef), and it was incompatible with PrivateLocation.
Thus, TypeScript would not allow you to use the reference type.