feat: inline scripts can automatically include snippets
This change makes it so that all snippets are imported to a single location.
Then, if a check's inline script references snippets, we generate local
aliases for them that import the snippet from the main location.

Another notable change is that the TypeScript AST walker now also looks for
require calls when collecting dependencies, because snippets are sort of
implicitly TypeScript files yet they may use require().

Legacy handlebars syntax is not supported and we throw an error if such use
is detected.