fix: create a separate file for each imported alert channel
Earlier, all alert channels of the same type were bundled into the same file.
Unfortunately when you did a per-resource import of the same type later, it
would clear everything else in the file and only include the newly generated
construct.

Now, each alert channels is in a separate file.