feat: use the new FileLoader in create-cli [sc-24412] (#1082)
* feat: use the new FileLoader in create-cli

We do not currently have a good way to share common code between the
workspace packages. For now, just copy the files until we have time for a
real solution.

It would have been too much trouble to bring over Bun and Deno detection,
so they were left out for now.

* fix: properly share the lock promise for unitialized loaders

While a legit fix, this is not an issue in our usage patterns because only
a single instance of the uninitialized loader was ever created.