feat: include jiti directly in create-cli because it's not a library (#1084)
The only supported usage of the create-cli is via the `npm create` command.
When used, the package gets installed to a global location and any
dependencies it tries to load are resolved from that global location.
Therefore even if the user adds jiti by themselves, it's not going to be
effective because the create-cli will not see it.

It makes sense to just include jiti outright in create-cli because it's
intended to be purely a CLI and not a library.