fix: default playwrightConfig was not being set for Browser/MultiStep checks
The issue was caused by the introduction of typed defaults in version 6.1.0.
However, it turns out that the type had never included `playwrightConfig`,
but it had still been working because all properties were overwritten blindly
regardless of whether they were present in the type or not.

This change makes BrowserChecks and MultiStepChecks use the appropriate
default playwrightConfig again. Tests were added to make sure it won't break
again.