feat: show an accurate description for Constructs in validation errors
Until now, validation errors would only show the underlying base type of the
construct. Many Constructs share the same base type, which makes it hard to
tell them apart.

Now, each Construct implements a `describe()` method which returns a unique
description for the Construct instance. It consists of the Construct class
name and the logicalId of the instance.