fix: always include pnpm-workspace.yaml for Playwright checks if present
Earlier, it was included because we added lockfiles using a glob pattern
which happened to match the workspace file as well. Now that we are only
including the lockfile we detect, we have to add the workspace file manually.

Done in an extremely simple manner because proper workspace support will
follow soon.